Move protocol-specific classes to mirai-core-qqandroid
