+++ mirai-core-qqandroid/src/commonMain/kotlin/net/mamoe/mirai/qqandroid/io/serialization/jce/JceInput.kt
@@ -11,7 +11,7 @@ package net.mamoe.mirai.qqandroid.io.serialization.jce
 
 import kotlinx.io.core.*
 import net.mamoe.mirai.qqandroid.io.serialization.JceCharset
-import net.mamoe.mirai.utils.io.readString
+import net.mamoe.mirai.qqandroid.utils.io.readString
 
 
 /**
